+++ apps/typegpu-docs/src/examples/rendering/heatmap/constants.ts
@@ -1,0 +1,6 @@
+import * as d from 'typegpu/data';
+
+export const target = d.vec3f(0, 0, 0);
+export const cameraInitialPos = d.vec4f(12, 7, 12, 1);
+export const futureNumOfReleases = 49;
+export const futureNumOfSamples = 49;
