+++ apps/typegpu-docs/src/examples/tests/tgsl-parsing-test/index.html
@@ -1,1 +1,1 @@
-<div class='result'>Wait for the tests to finish running...</div>;
+<div class="result">Wait for the tests to finish running...</div>;
